Usage integration, and cleanup
